refactor: Introduce shared workflow utilities and streamline UICustomizations for better maintainability and code reuse
