+++ frontend/micro-ui/web/micro-ui-internals/packages/modules/dristi/src/configs/UICustomizations.js
@@ -26,6 +26,7 @@ import { getAdvocates } from "../pages/citizen/FileCase/EfilingValidationUtils";
 import { OrderWorkflowState } from "../Utils/orderWorkflow";
 import { getFullName } from "../../../cases/src/utils/joinCaseUtils";
 import { CaseWorkflowState } from "../Utils/caseWorkflow";
+import { buildStandardInboxWorkflow } from "./uiCustomizationsWorkflowShared";
 
 export const getSelectedAdvocate = () => {
   try {
@@ -117,99 +118,25 @@ export const UICustomizations = {
   businessServiceMap,
   updatePayload: (applicationDetails, data, action, businessService) => {
     if (businessService === businessServiceMap.estimate) {
-      const workflow = {
-        comment: data?.comments,
-        documents: data?.documents?.map((document) => {
-          return {
-            documentType: action?.action + " DOC",
-            fileName: document?.[1]?.file?.name,
-            fileStoreId: document?.[1]?.fileStoreId?.fileStoreId,
-            documentUid: document?.[1]?.fileStoreId?.fileStoreId,
-            tenantId: document?.[1]?.fileStoreId?.tenantId,
-          };
-        }),
-        assignees: data?.assignees?.uuid ? [data?.assignees?.uuid] : null,
-        action: action?.action,
-      };
-      //filtering out the data
-      Object.keys(workflow).forEach((key, index) => {
-        if (!workflow[key] || workflow[key]?.length === 0) delete workflow[key];
-      });
-
       return {
         estimate: applicationDetails,
-        workflow,
+        workflow: buildStandardInboxWorkflow(data, action),
       };
     }
     if (businessService === businessServiceMap.contract) {
-      const workflow = {
-        comment: data?.comments,
-        documents: data?.documents?.map((document) => {
-          return {
-            documentType: action?.action + " DOC",
-            fileName: document?.[1]?.file?.name,
-            fileStoreId: document?.[1]?.fileStoreId?.fileStoreId,
-            documentUid: document?.[1]?.fileStoreId?.fileStoreId,
-            tenantId: document?.[1]?.fileStoreId?.tenantId,
-          };
-        }),
-        assignees: data?.assignees?.uuid ? [data?.assignees?.uuid] : null,
-        action: action?.action,
-      };
-      //filtering out the data
-      Object.keys(workflow).forEach((key, index) => {
-        if (!workflow[key] || workflow[key]?.length === 0) delete workflow[key];
-      });
-
       return {
         contract: applicationDetails,
-        workflow,
+        workflow: buildStandardInboxWorkflow(data, action),
       };
     }
     if (businessService === businessServiceMap?.["muster roll"]) {
-      const workflow = {
-        comment: data?.comments,
-        documents: data?.documents?.map((document) => {
-          return {
-            documentType: action?.action + " DOC",
-            fileName: document?.[1]?.file?.name,
-            fileStoreId: document?.[1]?.fileStoreId?.fileStoreId,
-            documentUid: document?.[1]?.fileStoreId?.fileStoreId,
-            tenantId: document?.[1]?.fileStoreId?.tenantId,
-          };
-        }),
-        assignees: data?.assignees?.uuid ? [data?.assignees?.uuid] : null,
-        action: action?.action,
-      };
-      //filtering out the data
-      Object.keys(workflow).forEach((key, index) => {
-        if (!workflow[key] || workflow[key]?.length === 0) delete workflow[key];
-      });
-
       return {
         musterRoll: applicationDetails,
-        workflow,
+        workflow: buildStandardInboxWorkflow(data, action),
       };
     }
     if (businessService === businessServiceMap?.["works.purchase"]) {
-      const workflow = {
-        comment: data?.comments,
-        documents: data?.documents?.map((document) => {
-          return {
-            documentType: action?.action + " DOC",
-            fileName: document?.[1]?.file?.name,
-            fileStoreId: document?.[1]?.fileStoreId?.fileStoreId,
-            documentUid: document?.[1]?.fileStoreId?.fileStoreId,
-            tenantId: document?.[1]?.fileStoreId?.tenantId,
-          };
-        }),
-        assignees: data?.assignees?.uuid ? [data?.assignees?.uuid] : null,
-        action: action?.action,
-      };
-      //filtering out the data
-      Object.keys(workflow).forEach((key, index) => {
-        if (!workflow[key] || workflow[key]?.length === 0) delete workflow[key];
-      });
+      const workflow = buildStandardInboxWorkflow(data, action);
 
       const additionalFieldsToSet = {
         projectId: applicationDetails.additionalDetails.projectId,
